Work on 64-bit versions and PredicateParser - PostgreSQL headers for all architectures are now stored separately. Earlier we used some of them and different pg_config.h files. Other files may be generated as well, though. - Wrote a wrapper for libpq-fe.h to choose the correct one. - Made some small changes to handle GCC 4.2's new warnings. - Changed some integer types to NSInteger and NSUInteger, since they have different sizes. - We now build for ppc64 and x86_64 in addition to ppc and i386. - We also use different compilers for them. - PredicateParser is now linked to CoreData to handle the CoreData specific expression type. - The universal binary creation script doesn't work currently. - Moved some build settings to xcconfig files.
git-svn-id: http://basetenframework.org/svn/trunk@1329 fe1aa653-7022-0410-af67-bafd91c7e665

+++ Sources/BXDatabaseContext.m
@@ -996,8 +996,8 @@ ModTypeToObject (enum BXModificationType value)
 @implementation BXDatabaseContext (Undoing)
 - (void) undoGroupWillClose: (NSNotification *) notification
 {
-	unsigned int groupingLevel = [mUndoManager groupingLevel];
-	unsigned int currentLevel = NSNotFound;
+	NSUInteger groupingLevel = [mUndoManager groupingLevel];
+	NSUInteger currentLevel = NSNotFound;
 	BOOL shouldEstablishSavepoint = NO;
 	while (NSNotFound != (currentLevel = [mUndoGroupingLevels lastIndex]))
 	{
@@ -1043,10 +1043,10 @@ ModTypeToObject (enum BXModificationType value)
 	}
 	else
 	{
-		int lastLevel = [mUndoGroupingLevels lastIndex];
+		NSUInteger lastLevel = [mUndoGroupingLevels lastIndex];
 		if (lastLevel != groupingLevel)
 		{
-			BXAssertValueReturn (NSNotFound == (unsigned) lastLevel || lastLevel < groupingLevel, NO, 
+			BXAssertValueReturn (NSNotFound == lastLevel || lastLevel < groupingLevel, NO, 
 								   @"Undo group level stack is corrupt.");
 			[mUndoGroupingLevels addIndex: groupingLevel];
 		}
